# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Reels App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateUser
      tags: [users]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

    delete:
      operationId: deleteUser
      tags: [users]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /users/{id}/follow:
    post:
      operationId: followUser
      tags: [users]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Followed

    delete:
      operationId: unfollowUser
      tags: [users]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Unfollowed
  /users/{id}/followers:
    get:
      operationId: getUserFollowers
      tags: [users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserPublic"
  /users/{id}/following:
    get:
      operationId: getUserFollowing
      tags: [users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserPublic"
  /users/{id}/reels:
    get:
      operationId: getUserReels
      tags: [users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reel"
  /users/{id}/posts:
    get:
      operationId: getUserPosts
      tags: [users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"

  # ── Reels ────────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Auth ───────────────────────────────────────────────────────────────────
...
    User:
      type: object
      required: [id, username, displayName, email, role, isActive, isSuspended, points, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        email:
          type: string
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        role:
          type: string
        isActive:
          type: boolean
        isSuspended:
          type: boolean
        points:
          type: integer
        createdAt:
          type: string
          format: date-time
    UserPublic:
      type: object
      required: [id, username, displayName]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
    UserProfile:
      type: object
      required: [id, username, displayName, followersCount, followingCount, reelsCount, postsCount]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        followersCount:
          type: integer
        followingCount:
          type: integer
        reelsCount:
          type: integer
        postsCount:
          type: integer
        isFollowing:
          type: boolean
    UserUpdate:
      type: object
      properties:
        displayName:
          type: string
        bio:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]

    # ── Reels ──────────────────────────────────────────────────────────────────
    Reel:
      type: object
      required: [id, userId, views, likesCount, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        user:
          $ref: "#/components/schemas/UserPublic"
        description:
          type: ["string", "null"]
        mediaUrl:
          type: ["string", "null"]
        thumbnailUrl:
          type: ["string", "null"]
        views:
          type: integer
        likesCount:
          type: integer
        music:
          type: ["string", "null"]
        likedByMe:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    Post:
      type: object
      required: [id, userId, content, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        user:
          $ref: "#/components/schemas/UserPublic"
        content:
          type: string
        mediaUrl:
          type: ["string", "null"]
        reactions:
          type: array
          items:
            $ref: "#/components/schemas/ReactionSummary"
        myReaction:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    ReactionSummary:
      type: object
      required: [emoji, count]
      properties:
        emoji:
          type: string
        count:
          type: integer

    # ── Store ──────────────────────────────────────────────────────────────────
...