# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Reels App API
 * OpenAPI spec version: 0.1.0
 */

export interface UserUpdate {
  displayName?: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
}